+++ miden-crypto/src/merkle/smt/full/mod.rs
@@ -15,6 +15,16 @@ mod proof;
 pub use proof::SmtProof;
 use winter_utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable};
 
+#[cfg(feature = "concurrent")]
+mod large;
+#[cfg(feature = "concurrent")]
+pub use large::{
+    LargeSmt, LargeSmtError, MemoryStorage, SmtStorage, StorageUpdateParts, StorageUpdates,
+    Subtree, SubtreeError,
+};
+#[cfg(feature = "rocksdb")]
+pub use large::{RocksDbConfig, RocksDbStorage};
+
 // Concurrent implementation
 #[cfg(feature = "concurrent")]
 mod concurrent;