+++ miden-crypto/src/merkle/smt/mod.rs
@@ -7,7 +7,14 @@ use super::{EmptySubtreeRoots, InnerNodeInfo, MerkleError, NodeIndex, SparseMerk
 use crate::{EMPTY_WORD, Felt, Map, Word, hash::rpo::Rpo256};
 
 mod full;
+#[cfg(feature = "concurrent")]
+pub use full::{
+    LargeSmt, LargeSmtError, MemoryStorage, SmtStorage, StorageUpdateParts, StorageUpdates,
+    Subtree, SubtreeError,
+};
 pub use full::{MAX_LEAF_ENTRIES, SMT_DEPTH, Smt, SmtLeaf, SmtLeafError, SmtProof, SmtProofError};
+#[cfg(feature = "rocksdb")]
+pub use full::{RocksDbConfig, RocksDbStorage};
 #[cfg(feature = "internal")]
 pub use full::{SubtreeLeaf, build_subtree_for_bench};