+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.17.0 (TBD)
 
+- Added `LargeSmt`, SMT backed by RocksDB ([#438](https://github.com/0xMiden/miden-crypto/pull/438)).
+
 ## 0.16.1 (2025-08-21)
 
 - Fix broken imports in CPU-specific `rescue` implementations (AVX2, SVE) ([#492](https://github.com/0xMiden/crypto/pull/492/)).